# openapi.yaml
openapi: 3.1.0
info:
  title: Office v2 Real-Time Telemetry & Control API
  version: 1.0.0
  description: |
    Kontrak API v1 untuk Office v2: Dunia isometrik 2.5D pixel-art kantor agen virtual Hermes.
    Backend menyajikan status kerja, aktivitas real-time, dan telemetri sistem secara 100% read-only
    terhadap runtime Hermes. Akses publik disanitasi menggunakan strict whitelist dual-projection,
    sedangkan kontrol operasional dan event kolektif dilindungi oleh autentikasi Founder.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://office.rifqisetiawan.my.id
    description: Production Server (Caddy Reverse Proxy)
  - url: https://office-staging.rifqisetiawan.my.id
    description: Staging Environment
security:
  - {}
paths:
...
  /api/v1/auth/login:
    post:
      summary: Login Founder
      description: |
        Mengautentikasi Founder menggunakan verifikasi hash Argon2. Jika kredensial valid,
        server mengembalikan cookie sesi office_session dengan atribut HttpOnly, Secure,
        dan SameSite=Strict.
        Wajib menyertakan header X-Office-Intent: 1 sebagai perlindungan mitigasi CSRF.
        Dibatasi maksimal 5 percobaan per 15 menit per alamat IP.
      operationId: loginFounder
      security:
        - IntentHeader: []
      parameters:
        - name: X-Office-Intent
          in: header
          required: true
          description: Header proteksi CSRF. Nilai harus bernilai "1".
          schema:
            type: string
            enum: ['1']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil, cookie sesi diterbitkan.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: office_session=eyJhbGciOi...; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Header X-Office-Intent tidak ada atau format payload salah.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Password Founder salah.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Header X-Office-Intent tidak ada atau nilai tidak valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Terlalu banyak percobaan login (maks. 5 kali per 15 menit).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      summary: Logout Founder
      description: |
        Menghapus cookie sesi Founder dan mengakhiri sesi administratif.
        Wajib menyertakan header X-Office-Intent: 1.
      operationId: logoutFounder
      security:
        - CookieAuth: []
          IntentHeader: []
      parameters:
        - name: X-Office-Intent
          in: header
          required: true
          description: Header proteksi CSRF. Nilai harus bernilai "1".
          schema:
            type: string
            enum: ['1']
      responses:
        '200':
          description: Logout berhasil, cookie sesi direset.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: office_session=; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '400':
          description: Header X-Office-Intent tidak ada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Sesi tidak valid atau telah berakhir.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Header X-Office-Intent tidak ada atau nilai tidak valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: office_session
      description: Signed session cookie untuk autentikasi Founder.
    IntentHeader:
      type: apiKey
      in: header
      name: X-Office-Intent
      description: Header anti-CSRF wajib bernilai "1" untuk seluruh mutasi POST.
  schemas:
...
    LoginRequest:
      type: object
      description: Kredensial autentikasi Founder.
      required:
        - password
      properties:
        password:
          type: string
          format: password
          description: Kata sandi Founder.
          example: SuperSecretFounderPassphrase2026!
    LoginResponse:
      type: object
      description: Respons keberhasilan autentikasi Founder.
      required:
        - success
        - role
        - expires_at
      properties:
        success:
          type: boolean
          example: true
        role:
          type: string
          example: founder
        expires_at:
          type: integer
          description: Unix timestamp detik berakhirnya masa aktif sesi.
          example: 1791633103
    LogoutResponse:
      type: object
      description: Respons pengakhiran sesi Founder.
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      description: Format standar representasi kesalahan API.
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Pesan deskripsi kesalahan yang mudah dipahami.
          example: Autentikasi Founder diperlukan untuk mengakses resource ini
        code:
          type: string
          description: Kode klasifikasi error mesin.
          example: UNAUTHORIZED
        details:
          type: [object, "null"]
          description: Detail konteks tambahan atas error jika ada.
          additionalProperties: true

    # SSE Event Payloads
...